requirements.txt
numpy==1.26.2
pandas==2.1.4
scikit_learn==1.3.2
scipy==1.11.4
gensim==4.3.0
nltk==3.9.1
openpyxl==3.1.5
networkx==3.4.2
matplotlib==3.10.1
plotly==6.0.1
UpSetPlot==0.9.0
seaborn==0.13.2
wordcloud==1.9.4

# AQua dependencies
adapter_transformers==3.2.1
huggingface-hub==0.19.4
tokenizers==0.13.3
# torch==1.12.1+cu113
torch
tqdm==4.66.1
